requirements.txt
Flask==3.0.3
Jinja2==3.1.4
yt-dlp==2025.06.09
gunicorn==22.0.0
redis==5.0.4
ffmpeg-python==0.2.0
python-dotenv==1.0.1
rq==1.16.2
rq-dashboard==0.7.1
